Split up ExportContainer Task in several sub-tasks.
