+++ exasol/slc/internal/tasks/export/cache_file_parameters.py
@@ -1,0 +1,6 @@
+import luigi
+
+
+class CacheFileParameters:
+    cache_file_path: str = luigi.Parameter()  # type: ignore
+    checksum_file_path: str = luigi.Parameter()  # type: ignore
